pyproject.toml
[project]
name = "datawarehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "django>=5.2.8",
    "django-debug-toolbar>=6.1.0",
    "django-filter>=25.2",
    "django-mptt>=0.18.0",
    "djangorestframework>=3.16.1",
    "faker>=38.2.0",
    "jdatetime>=5.2.0",
    "psycopg[binary]>=3.3.0",
    "python-decouple>=3.8",
]
